feat(local,builtins): restore everruns-local facade and promote InMemoryMessageRetriever
Two pieces of embedder-experience feedback from the 0.19 crate wave:

1. everruns-local was absorbed into everruns::local and its old 0.18.0 (which
   pinned everruns-host ^0.18) was yanked, leaving standalone dependants with no
   host-0.19 path. Reintroduce everruns-local 0.19.0 as a thin compatibility
   facade that re-exports everruns::local (via everruns with
   default-features=false, features=["local"]), so existing dependants compile
   against everruns-host 0.19 without pulling the incompatible 0.18 package. New
   code should depend on everruns directly with the local feature.

2. InMemoryMessageRetriever lived only in everruns-test-support (documented
   test-only) yet embedders need it in the production build path. Move it to
   everruns-builtins (next to the neutral MessageRetriever contract) as a
   supported production API; everruns-test-support re-exports it so existing test
   users are unchanged. builtins gains tokio "sync"; test-support takes builtins
   as a normal dependency for the re-export (no cycle).

Version bumps (published by the Crate Release workflow on merge): everruns-local
0.19.0 (new facade), everruns-builtins 0.18.1 -> 0.18.2 (additive), and
everruns-test-support 0.18.1 -> 0.18.2 (re-export). Pins synced; docs, doc-test,
publish-order, and strand-check guards pass.
