+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "crates/server",
     "crates/worker",
     "crates/everruns",
+    "crates/local",
     "crates/macros",
     "crates/capability",
     "crates/builtins",
@@ -174,12 +175,12 @@ futures-util = "0.3"
 # Everruns SDK
 everruns-sdk = "0.2.0"
 everruns-capability = { path = "crates/capability", version = "0.18.0", default-features = false }
-everruns-builtins = { path = "crates/builtins", version = "0.18.1" }
+everruns-builtins = { path = "crates/builtins", version = "0.18.2" }
 everruns-core = { path = "crates/core", version = "0.18.0" }
 everruns-engine = { path = "crates/engine", version = "0.18.0" }
 everruns-host = { path = "crates/host", version = "0.19.0" }
 everruns-llmsim = { path = "crates/drivers/llmsim", version = "0.18.1", default-features = false }
-everruns-test-support = { path = "crates/test-support", version = "0.18.1", default-features = false }
+everruns-test-support = { path = "crates/test-support", version = "0.18.2", default-features = false }
 everruns-platform = { path = "crates/platform", version = "0.18.1" }
 everruns-provider = { path = "crates/provider", version = "0.18.0", default-features = false }
 everruns-mcp = { path = "crates/mcp", version = "0.19.0" }
