+++ crates/local/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "everruns-local"
+version = "0.19.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+documentation = "https://docs.rs/everruns-local"
+description = "Compatibility facade re-exporting the Everruns local in-process backend (everruns::local) for everruns-host 0.19"
+readme = "README.md"
+keywords = ["everruns", "agent", "sqlite", "embedded", "storage"]
+categories = ["development-tools", "database"]
+
+[dependencies]
+# The local backend was absorbed into the `everruns` facade in 0.19; this crate
+# re-exports `everruns::local` so existing dependants keep compiling against
+# everruns-host 0.19 without the old, now-incompatible 0.18 package.
+everruns = { path = "../everruns", version = "0.18.1", default-features = false, features = ["local"] }
