+++ crates/test-support/src/lib.rs
@@ -73,7 +73,8 @@ pub use everruns_llmsim::{
     LlmSimConfig, LlmSimDriver, OnExhausted, ResponseConfig, SimError, SimToolCall, SimTurn,
     ToolCallConfig, ToolCallPattern,
 };
-pub use in_memory::{InMemoryEventEmitter, InMemoryMessageRetriever};
+pub use everruns_builtins::InMemoryMessageRetriever;
+pub use in_memory::InMemoryEventEmitter;
 #[cfg(all(feature = "sim", feature = "host"))]
 pub use in_memory_loop::{InMemoryAgenticLoop, InMemoryAgenticLoopBuilder};
